+++ gainsworth/cogs/gainsworth_clock.py
@@ -76,10 +76,10 @@ class GainsClock(commands.Cog):
         await self.client.wait_until_ready()
 
 
-def setup(client):
+async def setup(client):
     """
     This setup function must exist in every cog file and will ultimately have a
     nearly identical signature and logic to what you're seeing here.
     It's ultimately what loads the Cog into the bot.
     """
-    client.add_cog(GainsClock(client))
+    await client.add_cog(GainsClock(client))